move the `Object` class to a submodule
To avoid circular imports.
